backend/requirements.txt
flask==3.1.2
flask-cors==6.0.1
flask-wtf==1.2.2
qiskit==1.3.0
qiskit-aer==0.13.0
qiskit-ibm-runtime==0.21.2
amazon-braket-sdk==1.106.0
boto3==1.41.1
numpy==1.26.4
python-dotenv==1.2.1
flask-jwt-extended==4.6.0
flask-bcrypt==1.0.1
flask-sqlalchemy==3.1.1
stripe==12.0.0
gunicorn==23.0.0
